Created //rs/tests/ckbtc library from code in rs/tests/src/ckbtc/ . Not all the code in rs/tests/src/ckbtc was moved out. I only moved what I needed in order to move all the code out of rs/tests/src/nns_tests/bitcoin_set_config_proposal.rs to rs/tests/nns/bitcoin_set_config_proposal_test.rs. Perhaps, in a future commit, the rest of the code in rs/tests/src/ckbtc can be moved to rs/tests/ckbtc (or better yet, deleted completely!).
