+++ rs/tests/ckbtc/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "ic-tests-ckbtc"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+documentation.workspace = true
+
+[dependencies]
+canister-test                                = { path = "../../rust_canisters/canister_test" }
+dfn_candid                                   = { path = "../../rust_canisters/dfn_candid" }
+ic-base-types                                = { path = "../../types/base_types" }
+ic-canister-client                           = { path = "../../canister_client" }
+ic-ckbtc-kyt                                 = { path = "../../bitcoin/ckbtc/kyt" }
+ic-ckbtc-minter                              = { path = "../../bitcoin/ckbtc/minter" }
+ic-config                                    = { path = "../../config" }
+ic_consensus_threshold_sig_system_test_utils = { path = "../../tests/consensus/tecdsa/utils" }
+ic-icrc1-ledger                              = { path = "../../ledger_suite/icrc1/ledger" }
+ic-management-canister-types                 = { path = "../../types/management_canister_types" }
+ic-nervous-system-common-test-keys           = { path = "../../nervous_system/common/test_keys" }
+ic-nns-common                                = { path = "../../nns/common" }
+ic-nns-constants                             = { path = "../../nns/constants" }
+ic-nns-governance-api                        = { path = "../../nns/governance/api" }
+ic-nns-test-utils                            = { path = "../../nns/test_utils" }
+ic-registry-subnet-features                  = { path = "../../registry/subnet_features" }
+ic-registry-subnet-type                      = { path = "../../registry/subnet_type" }
+ic-system-test-driver                        = { path = "../../tests/driver" }
+ic-types                                     = { path = "../../types/types" }
+ic-types-test-utils                          = { path = "../../types/types_test_utils" }
+icp-ledger                                   = { path = "../../ledger_suite/icp" }
+registry-canister                            = { path = "../../registry/canister" }
+
+candid                                       = { workspace = true }
+ic-agent                                     = { workspace = true }
+ic-btc-interface                             = { workspace = true }
+slog                                         = { workspace = true }
