+++ rs/tests/src/ckbtc/mod.rs
@@ -1,4 +1,6 @@
 pub mod agent;
-pub mod btc_config;
+pub mod btc_config {
+    pub use super::lib::btc_config as config;
+}
 pub mod lib;
 pub mod minter;
